REFACTORY tests
