+++ test/__init__.py
@@ -1,0 +1,3 @@
+GATEWAY_URL = "http://localhost:8080"
+PORTS = ["3000", "3001", "3002"]
+IDS = ["Bob", "Alice", "James"]
